SimCLR adjusted and working, refactor the rest; ConRecon next up
