+++ experiments/ssl/backbones/resnet50x2.yaml
@@ -1,9 +1,9 @@
-CONSTANTS#version: "lr=dynamic_global_grid_size=[4,8,8]_ResNet50x2"
-CONSTANTS#HIGH_RES_GRID_SIZE: [4, 8, 8]
+CONSTANTS#backbone_name: ResNet50x2
+# Used by VicRegL. For the specified ResNet50x2 architecture, the output features will be the input shape divided by 32.
+CONSTANTS#HIGH_RES_GRID_SIZE: "$[x // 32 for x in @CONSTANTS#HIGH_RES_SIZE]"
 
-system#model#num_ftrs: 4096
 system#model#backbone: 
-      _target_: project.models.resnet.ResNet
+      _target_: project.models.backbones.resnet.ResNet
       block: "bottleneck"
       layers: [3, 4, 6, 3]
       block_inplanes: [64, 128, 256, 512]
