+++ experiments/ssl/backbones/segresnetds.yaml
@@ -1,11 +1,12 @@
-CONSTANTS#version: "lr=dynamic_global_grid_size=[8,16,16]_SegResNetDS"
-CONSTANTS#HIGH_RES_GRID_SIZE: [8, 16, 16]
+CONSTANTS#backbone_name: SegResNetDS
+# Used by VicRegL. For the specified SegResNet architecture, the output features will be the input shape divided by 16.
+CONSTANTS#HIGH_RES_GRID_SIZE: "$[x // 16 for x in @CONSTANTS#HIGH_RES_SIZE]"
 
-system#model#num_ftrs: 512
 system#model#backbone: 
-    _target_: project.models.segresnetds.SegResNetDS
+    _target_: monai.networks.nets.SegResNetDS
     spatial_dims: 3
     in_channels: 1
-    out_channels: 1  # Doesn't matter, not training the decoder
+    out_channels: 1
     init_filters: 32
     blocks_down: [1, 2, 2, 4, 4]
+    dsdepth: 4
