+++ experiments/ssl/backbones/segresnetds_w_embedding.yaml
@@ -1,9 +1,7 @@
-CONSTANTS#version: "lr=dynamic_global_grid_size=[8,16,16]_SegResNetDS"
-CONSTANTS#HIGH_RES_GRID_SIZE: [8, 16, 16]
+CONSTANTS#backbone_name: SegResNetDS
 
-system#model#num_ftrs: 512
 system#model#backbone: 
-    _target_: project.models.segresnetds_w_embedding.SegResNetDSwEmbedding
+    _target_: project.models.backbones.segresnetds_w_embedding.SegResNetDSwEmbedding
     spatial_dims: 3
     in_channels: 1
     out_channels: 1  # Doesn't matter, not training the decoder
