+++ loss/intrasample_ntxent_loss.py
@@ -22,58 +22,50 @@ class IntraSampleNTXEntLoss(nn.Module):
         self.eps = 1e-8
 
         if abs(self.temperature) < self.eps:
-            raise ValueError(
-                "Illegal temperature: abs({}) < 1e-8".format(self.temperature)
-            )
-        
+            raise ValueError("Illegal temperature: abs({}) < 1e-8".format(self.temperature))
 
     def forward(self, input):
         """
         The forward pass of the IntraSampleNTXEntLoss.
 
-
         Args:
-            out (tuple of tuple of Tensors): A tuple of input data. 
-            Each element of the tuple represents a sample drawn from an image (dataset item)
-            Each sample is a tuple of Tensors representing different views of the same image.
-            Each view is a Tensor of shape (batch_size, channels, height, width).
-
+            input (list of list of Tensors): The input data in the following format:
+                List [views], List [crops], Tensor [BC(D)HW].
+                Each batch index represents a different image.
         Returns:
             float: Contrastive Cross Entropy Loss value.
         """
-
-        device = input[-1][-1].device
+        # Dimensions: [views, crops, batch size, channels, (depth), height, width]
         batch_size = input[-1][-1].shape[0]
+        num_crops = len(input[-1])
 
-        overall_loss = 0
+        loss = 0
         for b in range(batch_size):
             per_sample_loss = 0
-            for s_i, sample in enumerate(input):
-                pos0 = sample[0][b].unsqueeze(0)
-                pos0 = nn.functional.normalize(pos0, dim=1)
+            for i in range(num_crops):
+                # Positives: both views (0 and 1), same crop (i), from the same image (b - batch index)
+                positive_0 = input[0][i][b].unsqueeze(0)
+                positive_1 = input[1][i][b].unsqueeze(0)
 
-                pos1 = sample[1][b].unsqueeze(0)
-                pos1 = nn.functional.normalize(pos1, dim=1)
-                
-                negatives = [input[s_j][0][b] for s_j, _ in enumerate(input) if s_j != s_i]
-                negatives.extend([input[s_j][1][b] for s_j, _  in enumerate(input) if s_j != s_i])
-
+                # Negatives: both views (0 and 1), all other crops (j != i), from the same image (b - batch index)
+                negatives = []
+                for j in range(num_crops):
+                    if j != i:
+                        negatives.append(input[0][j][b])
+                        negatives.append(input[1][j][b])
                 negatives = torch.stack(negatives, dim=0)
+
+                positive_0 = nn.functional.normalize(positive_0, dim=1)
+                positive_1 = nn.functional.normalize(positive_1, dim=1)    
                 negatives = nn.functional.normalize(negatives, dim=1)
 
-                sim_pos = torch.einsum("nc,nc->n", [pos0, pos1]).unsqueeze(-1)
-                sim_neg = torch.einsum("nc,kc->nk", [pos0, negatives])
+                sim_pos = torch.einsum("nc,nc->n", positive_0, positive_1).unsqueeze(-1)
+                sim_neg = torch.einsum("nc,kc->nk", positive_0, negatives)
 
                 logits = torch.cat([sim_pos, sim_neg], dim=1) / self.temperature
-                labels = torch.zeros(logits.shape[0], dtype=torch.long, device=device)
+                labels = torch.zeros(logits.shape[0], dtype=torch.long, device=logits.device)
 
-                loss = self.cross_entropy(logits, labels)
-
-                per_sample_loss += loss
-
-            per_sample_loss /= len(input)
-            overall_loss += per_sample_loss
-
-        overall_loss /= batch_size
-
-        return overall_loss
+                per_sample_loss += self.cross_entropy(logits, labels)
+            loss += per_sample_loss / num_crops
+        loss /= batch_size
+        return loss
